feat(status-page): add React frontend (#28)
## Background / 背景

status page（`bons8i.hagaspa.com` に公開予定）の Day 2 = フロントエンド実装。PR #27 の Rust
BFF が返す `/api/status` を可視化する React UI を追加する。設計:
`obsidian/projects/k8s/pi/status-page-design.md`。

## Changes / 変更内容

- `web/status-page/frontend/`: Vite + React 19 + TypeScript の CSR を新規追加
- 健康状態バッジ（operational / degraded / unknown / API 不達）、発火中アラート一覧、Cluster
メトリクスカード 5 枚、Alert history（GitHub issue 統計）カード 3 枚、bons8i へのリンク
- **API 型は ts-rs 生成型（`src/generated/`）のみを import** — 手書きの型定義ゼロ、Rust
が単一ソース
  - 60 秒ポーリング（BFF のクラスタキャッシュ TTL と同周期）。fetch 失敗は unreachable バッジに劣化
  - dev proxy: `/api` → `127.0.0.1:18080`（`VITE_API_TARGET` で上書き可）
  - スタイルは素の CSS。**薄い緑（bons8i のイメージカラー）を敷いた単一ライトテーマ**に白カード
- `server/`: `IssueStats` の `u64` を `u32` に変更。ts-rs は u64 を `bigint`
に写すが `JSON.parse` の実行時値は `number` であり、生成型と実態が食い違うため（u32 → number
で一致）。`IssueStats.ts` は再生成
- ビルド成果物: JS 194KB / gzip 61KB

## Impact scope / 影響範囲

- `web/status-page/` のみ。クラスタ上のリソースへの変更なし（デプロイは Day 4 の別 PR）

## Testing / 動作確認

- [x] `npm run build`（tsc + vite build）通過
- [x] `cargo test` 通過（u32 化後の型再生成を確認）
- [x] BFF + Vite dev proxy 経由で実データのブラウザ表示を確認（All Systems Operational /
63.6°C / 28 pods / MTTR 1.1h、スクリーンショットは
`obsidian/projects/k8s/pi/day24-status-page-frontend.png`）
- [x] 見た目は本人確認済み（白カード + 薄緑背景で確定）

+++ web/status-page/server/src/upstream.rs
@@ -149,7 +149,7 @@ pub async fn fetch_issue_stats(
         .filter(|i| i.pull_request.is_none())
         .collect();
 
-    let open_count = issues.iter().filter(|i| i.state == "open").count() as u64;
+    let open_count = issues.iter().filter(|i| i.state == "open").count() as u32;
     let closed_30d: Vec<_> = issues
         .iter()
         .filter(|i| i.closed_at.is_some_and(|c| c > cutoff))
@@ -164,7 +164,7 @@ pub async fn fetch_issue_stats(
 
     Ok(IssueStats {
         open_count,
-        closed_count_30d: closed_30d.len() as u64,
+        closed_count_30d: closed_30d.len() as u32,
         avg_hours_to_close_30d: avg_hours,
     })
 }
